Peephole - UNARY_NOT POP_JUMP_IF_FALSE  -->  POP_JUMP_IF_TRUE
Summary: Implements the first peep hole optimization.  These are all basically just ports from CPython's peephole.c with the addition of test_peepholer.py tests which run the compiler with the peephole optimizer on.  The optimizer is not yet enabled by default because there's no way to get an apples to apples code gen comparison with it in a partially implemented state.

Test Plan: ./python -m test.test_compiler