Fix perf problem w/ huge const pools exposed by test_folding_of_tuples_of_constants
Summary:
One tuple folding test creates a huge number of constants and when running the test suite it seems as if it just hangs.  This is due to a performance problem where it's doing a linear scan of the constant pool.  This replaces it with a dictionary lookup that includes the type.

This is probably still not 100% correct and I'll follow up with additional test cases and fixes as appropriate in a later diff.  But it does match the behavior of the current compiler which is just using type equality as well.

Test Plan: ./python -m test.test_compiler