Don't create intermediate code object
Summary:
When a code object is constructed in Python it does a pass of running
through and interning all of the strings which can be used as identifiers.
When it does this it will scan (and mutate!) not only the tuples, but
it'll also scan into frozen sets.  To do this it creates a tuple of
the elements in the frozen set, interns the strings in the tuple, and
then builds a new frozen set.

Ultimately this results in a newly created frozen set that can have
a different order than the original set. A simple demonstration of this is to repeatedly run this program:

```
from types import CodeType
def add(a, b): return a + b
s =  frozenset([add("gi", "_running"),
                        add("gi", "_frame"),
                        add("gi", "_code"),
                        add("gi", "_yieldfrom"),
                        add("cr", "_running"),
                        add("cr", "_frame"),
                        add("cr", "_code"),
                        add("cr", "_await")])

c = CodeType(0, 0, 1, 1, 1, b'', (s, ), (), (), '', '', 1, b'')
print(repr(c.co_consts[0]) == repr(s))
```

Sometimes this will print True, other times it will print False, depending on the current string hash seed.  It also captures all of the necessary pieces: the strings can't yet already have been interned (hence the weird add), otherwise the frozenset doesn't get re-built.  And we have to be unlucky enough that when pulling the values into a tuple out of the frozenset that we get a different order than we initially went into the frozen set with.

Currently we hit this behavior because we first construct the code object, then we optimize it, and then we re-construct it.  Going through that 1st time the strings get interned and the order of the sets can be perturbed.

So now we don't construct the code object until after we've optimized it.  The peephole optimizer just works on the portions of the code object it requires: the byte code, the lnotab (to update the line information), and the consts (for the const folding).  That carefully matches CPython's hashing behavior so that we get 100% identical code objects back.

Note this shouldn't have any practical difference in properly constructed programs.  But it makes it easier for us to assert that we have an identical compiler, and people have been known to introduce various dependencies on hashing.

Test Plan: ./python -m test.test_compiler