__class__ fixes
Summary:
`__class__` is a weird and special snowflake.  It gets defined at the end of a class body, and is automatically closed over when you refer to super.  But it's the only class-level variable that you can close over - otherwise variables declared need to go through MyClass.my_var to be accessed.

The existing Python symbol binder deals with this via a bunch of special cases around the usage of `__class__`.  This attempts to reduce some of those special cases and fixes up several places where the handling wasn't correct.  It doesn't really make the Python symbol table look like the CPython one - the CPython one just has a couple of places where it defines and clears the __class__ variable.  But it does seem to get us the same behavior without a total rewrite of the Python implementation.

**get_free_vars**:
It wasn't entirely obvious what this was doing with it's weird asserts.  But if a function isn't nested then it should only have 1 variable which it's going to close over, which is __class__.  This happens in the normal case where you have a function inside of a class statement that accesses super().  This just kind of simplifies that and makes it more obvious we just are dealing with __class__ there.

**handle_chidlren**:
This is trying to get the free variables that the children access, and push them up as free variables into the parent.  If now split this into the nested/non-nested cases.  For nested it remains the same, but for the non-nested cases we recognize we only care about __class__ and add it to our frees and make sure it's marked as a free in the child.  The forcing it as a free in the child is necessary because it's not recognized as being defined in the parent in a normal way.

**add_frees**:
This was just too narrowly tailored and wasn't accounting for additional levels of nesting that can occur.

**Nonlocal**:
I'm not really sure what the old logic was attempting to get at here...  Going down the "add_free" path has a bunch of weird side effects that are largely based upon otherwise uninitialized state.  The additional checks for __class__ also seemed like a weird place for these to go when there was already logic to deal w/ fixing these up after analyzing the child scopes.  So this was easy to greatly simplify because nonlocal is always going to create a free variable as it cannot be used for any other purpose - for example you can't use it to access global variables, you have to use global for that.

Test Plan: ./python -m test.test_compiler