Remove assign args in symbol visitor
Summary:
The symbol visitor is explicitly passing around whether or not something should be treated as an assignment in order to special case certain assignments.  Some of the nodes are actually missing it though (await, yield, and yield from).  But Python's assignment context is good enough to track these.  So this just gets rid of the parameter across the board rather than adding it to the places where it's missing.

Also adds test cases that make sure assignments get properly tracked.

Test Plan: ./python -m test.test_compiler