Clean up code gen constructors
Summary:
This is the next step - a small cleanup of the code gen constructors.  This is getting rid of the weird alternative to super() usage and makes the constructors do a lot less work.  Now instead of doing the code gen themselves, we do the work in normal functions.  This also exposes that the interactive code generator is basically totally busted and fixes it up as well.

visitDiscard is probably an unrelated way to how interactive code was handled before, with an explicit node in the tree instead of having interactive mode just pop these.  That node no longer exists, so the visitor for it is removed.

Test Plan: ./python -m test.test_compiler