Treat peephole enabling/disabling as a CodeGenerator
Summary: This switches from using a flag which specifies whether or not the peep hole optimizer is enabled to using a different CodeGenerator which enables it / disables it.  I'm not a 100% convinced this is the best way to handle the peephole optimizer, but it's a weird piece of ugliness already.  And I want to avoid having it be some random orthogonal switch which needs to exist across multiple code generation strategies as well.  So this just sets the stage to say that you pick a CodeGenerator type and it'll encompass all the configuration for the compiler.

Test Plan: ./python -m test.test_compiler