Expose make_compiler API which is like compile() but takes a code generator
Summary: Adds a make_compiler API which allows constructing a compiler w/ a given code generator and gets rid of the AbstractCompileMode stuff.  Our code gen for different modes is now identical - it only varies based upon the types that are present.  So when compiling something for interactive mode there'll be an Interactive node, and likewise there's a Module node.  So all of this stuff is an unnecessary abstraction.

Test Plan: ./python -m test.test_compiler