There can be only 1: don't use a list for prev/next
Summary: This is just a minor cleanup of something that has bugged me about the compiler for a while - for some reason they use a list of blocks for the next/prev block.  But they expect the list to only have at maximum 1 item.  So now we just store the blocks directly and init them to None.

Test Plan: ./python -m test.test_compiler