Adds ast rewriter
Summary:
Python 3.7 starts doing various optimizations at the AST level, where it actually makes sense.  To do this we need the ability to re-write the ASTs.  This adds an ASTRewriter class which supports having the visit() methods return the new nodes to be re-written.  Child nodes which aren't modified remain in the tree, and the spine of the AST is rewritten as we walk back up the AST.

Even though the Python ASTs are mutable this supports performing the re-writes w/o modifying the AST in place.

Test Plan: ./python -m test.test_compiler