Fix constant comparison
Summary:
This was a potential bug I had noted long ago when previously improving the code gen to account for the type of objects so that 0.0 doesn't get turned into 0 even though they compare equal.

In addition to different times which represent the same numerical value you can have values which are the same type and not equal - 0.0 and -0.0 (and likewise (0.0, ) and (-0.0, )).  I don't think this was actually an issue before because -0.0 would be produced as a constant only by the peephole optimizer, and so the code generator would never have any -0.0's const's.  But now the optimizer runs first and we can end up with both a 0.0 and a -0.0 in the constant pool.

This makes the key which we use account for these differences as well.

Test Plan: ./python -m test.test_compiler